third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/styles-4/do-not-rebuild-styles-on-every-change-expected.txt

Tests show that ssp isn't rebuild on every dom mutation


Running: setupTest

Running: testModifySibling
Requested StyleSidebarPane update

Running: testModifySiblingChild

Running: testModifyParent
Requested StyleSidebarPane update

Running: testModifyChild
